switch to u16 buffer, use core1 callback to queue next scanline

+++ rust-eg-fb/rust-eg-fb.h
@@ -15,4 +15,4 @@ extern struct framebuffer FRAME;
 
 void framebuffer_draw(void);
 
-uint8_t *framebuffer_get_raw_pointer(void);
+uint16_t *framebuffer_get_raw_pointer(void);
